refactor(store): extract ConflictPredicate to shared conflict module
